requirements.txt
pandas
selenium
python-dotenv
snowflake-connector-python

